refactor(core): split shortcut installation by platform
